backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
langchain-chroma
pydantic
python-dotenv
httpx
